+++ grid-finder/include/Pixel.hpp
@@ -1,20 +1,20 @@
 #pragma once
 
-#include <cstddef>
+#include <uint.hpp>
 #include <ostream>
 
 struct Pixel {
-    constexpr Pixel(size_t x, size_t y) : x(x), y(y) {}
+    constexpr Pixel(uint x, uint y) : x(x), y(y) {}
     constexpr Pixel() : x(-1), y(-1) {}
-    size_t x, y;
-    bool isValid() const { return x != (size_t) -1 && y != (size_t) -1; }
+    uint x, y;
+    bool isValid() const { return x != (uint) -1 && y != (uint) -1; }
     bool operator==(Pixel other) const {
         return this->x == other.x && this->y == other.y;
     }
     bool operator!=(Pixel other) const {
         return this->x != other.x || this->y != other.y;
     }
-    bool inRange(size_t w, size_t h) const {
+    bool inRange(uint w, uint h) const {
         // Unsigned comparison also checks for >= 0
         return x < w && y < h;
     }
