+++ grid-finder/test/test-CenterPointOutLineIterator.cpp
@@ -6,18 +6,18 @@ using std::vector;
 
 TEST(CenterPointOutLineIterator, CenterPointOutLineIteratorOdd) {
     CenterPointOutLineIterator c(7);
-    vector<size_t> result;
+    vector<uint> result;
     while (c.hasNext())
         result.push_back(c.next());
-    vector<size_t> expect = {3, 4, 2, 5, 1, 6, 0};
+    vector<uint> expect = {3, 4, 2, 5, 1, 6, 0};
     ASSERT_EQ(result, expect);
 }
 
 TEST(CenterPointOutLineIterator, CenterPointOutLineIteratorEven) {
     CenterPointOutLineIterator c(6);
-    vector<size_t> result;
+    vector<uint> result;
     while (c.hasNext())
         result.push_back(c.next());
-    vector<size_t> expect = {2, 3, 1, 4, 0, 5};
+    vector<uint> expect = {2, 3, 1, 4, 0, 5};
     ASSERT_EQ(result, expect);
 }
